requirements.txt
Flask
newsapi-python

